refactor: move generateUniqueId to dom utilities
