+++ packages/dom/src/core/plugins/accessibility/utilities.ts
@@ -12,12 +12,3 @@ export function isFocusable(element: Element) {
     element instanceof window.HTMLAreaElement
   );
 }
-
-const ids: Record<string, number> = {};
-
-export function generateUniqueId(prefix: string) {
-  const id = ids[prefix] == null ? 0 : ids[prefix] + 1;
-  ids[prefix] = id;
-
-  return `${prefix}-${id}`;
-}
