+++ packages/dom/src/utilities/misc/generateUniqueId.ts
@@ -1,0 +1,8 @@
+const ids: Record<string, number> = {};
+
+export function generateUniqueId(prefix: string) {
+  const id = ids[prefix] == null ? 0 : ids[prefix] + 1;
+  ids[prefix] = id;
+
+  return `${prefix}-${id}`;
+}
